% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{connect_db}
\alias{connect_db}
\title{Establish a connection to the PostgreSQL database}
\usage{
connect_db(
  dbname,
  host = "localhost",
  port = 5432,
  user = NULL,
  password = NULL
)
}
\arguments{
\item{dbname}{The name of the database}

\item{host}{The database server host}

\item{port}{The port number}

\item{user}{The username for the database}

\item{password}{The password for the database}
}
\value{
A database connection object
}
\description{
This function creates a connection to a PostgreSQL database using the provided parameters.
}
