% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{execute_query}
\alias{execute_query}
\title{Execute a SQL query}
\usage{
execute_query(query, params = NULL, fetch = TRUE)
}
\arguments{
\item{query}{A string containing the SQL query to execute}

\item{params}{A list of parameters to be used in the query (for parameterized queries)}

\item{fetch}{Logical, if TRUE (default) fetches and returns the results}
}
\value{
If fetch is TRUE, returns a data frame with the query results. Otherwise, returns NULL.
}
\description{
This function executes a given SQL query on the connected database.
}
