% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{fetch_table}
\alias{fetch_table}
\title{Fetch all data from a table}
\usage{
fetch_table(table_name, schema = "public")
}
\arguments{
\item{table_name}{The name of the table to fetch data from}

\item{schema}{The schema name (optional, defaults to public)}
}
\value{
A data frame containing all rows from the specified table
}
\description{
This function retrieves all rows from a specified table.
}
