% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{get_connection}
\alias{get_connection}
\title{Retrieve the current database connection}
\usage{
get_connection()
}
\value{
A database connection object or NULL if no connection is active
}
\description{
This function returns the current database connection object.
}
